Move reg_read_long to x86 and arm64 modules
