+++ crates/unicorn/src/lib.rs
@@ -628,54 +628,6 @@ impl<'a, D, A: UcArch> Unicorn<'a, D, A> {
         }
     }
 
-    /// Read 128, 256 or 512 bit register value into heap allocated byte array.
-    ///
-    /// This adds safe support for registers >64 bit (GDTR/IDTR, XMM, YMM, ZMM, ST (x86); Q, V
-    /// (arm64)).
-    pub fn reg_read_long(&self, reg: A::Reg) -> Result<Box<[u8]>, uc_error> {
-        let curr_reg_id = reg.id();
-        let curr_arch = self.get_arch();
-
-        let value_size = match curr_arch {
-            #[cfg(feature = "arch_x86")]
-            Arch::X86 => Self::value_size_x86(curr_reg_id)?,
-            #[cfg(feature = "arch_arm")]
-            Arch::ARM64 => Self::value_size_arm64(curr_reg_id)?,
-            _ => Err(uc_error::ARCH)?,
-        };
-        let mut value = vec![0; value_size];
-        unsafe { uc_reg_read(self.get_handle(), curr_reg_id, value.as_mut_ptr().cast()) }
-            .and_then(|| Ok(value.into_boxed_slice()))
-    }
-
-    #[cfg(feature = "arch_arm")]
-    fn value_size_arm64(curr_reg_id: i32) -> Result<usize, uc_error> {
-        match curr_reg_id {
-            r if (RegisterARM64::Q0 as i32..=RegisterARM64::Q31 as i32).contains(&r)
-                || (RegisterARM64::V0 as i32..=RegisterARM64::V31 as i32).contains(&r) =>
-            {
-                Ok(16)
-            }
-            _ => Err(uc_error::ARG),
-        }
-    }
-
-    #[cfg(feature = "arch_x86")]
-    fn value_size_x86(curr_reg_id: i32) -> Result<usize, uc_error> {
-        match curr_reg_id {
-            r if (RegisterX86::XMM0 as i32..=RegisterX86::XMM31 as i32).contains(&r) => Ok(16),
-            r if (RegisterX86::YMM0 as i32..=RegisterX86::YMM31 as i32).contains(&r) => Ok(32),
-            r if (RegisterX86::ZMM0 as i32..=RegisterX86::ZMM31 as i32).contains(&r) => Ok(64),
-            r if r == RegisterX86::GDTR as i32
-                || r == RegisterX86::IDTR as i32
-                || (RegisterX86::ST0 as i32..=RegisterX86::ST7 as i32).contains(&r) =>
-            {
-                Ok(10)
-            }
-            _ => Err(uc_error::ARG),
-        }
-    }
-
     /// Read a signed 32-bit value from a register.
     pub fn reg_read_i32(&self, reg: A::Reg) -> i32 {
         let mut value = 0;
